JavaScript/FrameWorks/VueJs/Vue2/Projects/User_Profile_App/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
